+++ QuickFIXn/Message/Group.cs
@@ -63,6 +63,10 @@ namespace QuickFix
         /// </summary>
         public int Delim { get; }
 
+        /// <summary>
+        /// Creates a FIX (ish) string representation of this FieldMap (does not change the object state)
+        /// </summary>
+        /// <returns></returns>
         public override string CalculateString() {
             return base.CalculateString(new StringBuilder(), FieldOrder ?? new[] { Delim });
         }