+++ RELEASE_NOTES.md
@@ -22,6 +22,8 @@ What's New
 * #878 - corrections to tag 45 "Side" in various DDs (gbirchmeier) - most people won't notice, easy fix if they do
      * fix typo in FIX50 and FIX50SP1: `CROSS_SHORT_EXXMPT` fixed to `CROSS_SHORT_EXEMPT`
      * correction in FIX41 and FIX42: `D` to `UNDISCLOSED`
+* #863 - Change Message.ToString() to not alter object state anymore. (gbirchmeier)
+         Use new function Message.ConstructString() if you need BodyLength/CheckSum to be updated.
 
 **Non-breaking changes**
 * #864 - when multiple threads race to init DefaultMessageFactory,